+++ userscript/NinjaIOUtils.js
@@ -2517,7 +2517,6 @@
       App.Layer.emit(Layer.Events.HIDE_MENU);
       app.onResize();
     });
-    utilsButton.icon.scale.x = utilsButton.icon.scale.y = 0.7;
     menu.container.addChild(utilsButton);
     class ModsMenu extends Feature {
       ox = 40;
@@ -2673,18 +2672,17 @@
         if (mod.isInstalled() && mod.config) {
           const button = new Button("settings");
           button.scale.x = button.scale.y = 0.8;
-          button.setText("");
+          button.setText("  ");
           button.x = pl -= button.width + 8;
           button.y = pt;
-          button.width = button.height;
           button.addListener(Button.BUTTON_RELEASED, () => {
             this.showConfig(mod);
           });
           const ico = new PIXI.Sprite(App.CombinedTextures["menu_icon_settings"]);
-          ico.width = (button.width - 4) * button.scale.y;
-          ico.height = (button.height - 4) * button.scale.x;
-          ico.x = button.width / -3.5;
-          ico.y = button.height / 8;
+          ico.width = button.width;
+          ico.height = button.height;
+          ico.x = button.width * -0.2;
+          ico.y = button.height * 0.1;
           button.addChild(ico);
           container.addChild(button);
         }